Resolve issue #42
(1) Implements search feature. When a valid username is entered in the search bar, you are redirected to their timeline (skeleton page) if you are friends. If you are not friends, takes you to a separate page indicating you must be friends to view their timeline.
(2) Implements a basic skeleton for the friends page, with non-functional buttons for modifying relationships with friends and potential friends.
(3) Reworked Share in the database (renamed to Repost) so that the behavior will function more like Twitter retweets. If a user reposts a post, it will show up in their timeline. Friends viewing their timeline will see the reposted post.